+++ CHANGELOG.md
@@ -8,12 +8,12 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
-## [1.2.0] - 2025-04-08
+## [1.2.0] - 2025-04-??
 
 ### Added
 
 - [platform]: Added Robot Vacuum Cleaner device (supported by SmartThings, Alexa, Home Assistant and Apple Home).
-- [platform]: Added Laundry Washer deice (supported by SmartThings, Alexa and Home Assistant).
+- [platform]: Added Laundry Washer device (supported by SmartThings, Alexa and Home Assistant).
 - [platform]: Added Dishwasher device (supported by SmartThings, Alexa and Home Assistant).
 - [platform]: Added Temperature Controlled Cabinet Cooler device (supported by SmartThings, Alexa and Home Assistant).
 - [platform]: Added Temperature Controlled Cabinet Heater device (supported by SmartThings, Alexa and Home Assistant).